Refactor Header component and remove TikTokIcon
